improvement of cuckoo search algorithm
Former-commit-id: c480896d80cdd0808ce01d22f97d54717530a26e

+++ org.getalp.lexsema-wsd/org.getalp.lexsema-wsd-core/src/main/java/org/getalp/lexsema/wsd/configuration/ContinuousConfiguration.java
@@ -23,7 +23,7 @@ public class ContinuousConfiguration implements Configuration
         setRandomSenses();
     }
 
-    public ContinuousConfiguration(Document d, double[] senses)
+    public ContinuousConfiguration(Document d, int[] senses)
     {
         document = d;
         size = d.size();
